# docs/common/reference/api/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/users/stats": {
      "get": {
        "operationId": "UserController_getStats",
        "parameters": [],
        "responses": {
          "200": {
            "description": "통계 조회 성공"
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "관리자 대시보드 통계",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/profile": {
      "get": {
        "operationId": "UserController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiSuccess"
                    },
                    {
                      "type": "object",
                      "required": [
                        "success",
                        "data"
                      ],
                      "properties": {
                        "success": {
                          "type": "boolean",
                          "enum": [
                            true
                          ],
                          "example": true
                        },
                        "data": {
                          "$ref": "#/components/schemas/UserProfileDto"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "500": {
            "description": "서버 오류",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "내 프로필",
        "tags": [
          "users"
        ]
      },
      "put": {
        "operationId": "UserController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateOwnProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "프로필 수정 성공"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "내 계정 프로필 수정",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users": {
      "get": {
        "operationId": "UserController_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "페이지 번호",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "페이지당 항목 수",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "search",
            "required": false,
            "in": "query",
            "description": "검색어 (이름, 로그인ID, 이메일)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "roleCode",
            "required": false,
            "in": "query",
            "description": "역할 코드 필터",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 목록 조회 성공"
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "사용자 목록 (관리자)",
        "tags": [
          "users"
        ]
      },
      "post": {
        "operationId": "UserController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "사용자 생성 성공"
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "사용자 생성 (관리자)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}": {
      "put": {
        "operationId": "UserController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "사용자 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "사용자 수정 성공"
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "사용자 수정 (관리자)",
        "tags": [
          "users"
        ]
      },
      "delete": {
        "operationId": "UserController_deactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "사용자 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 비활성화 성공"
          },
          "403": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "사용자 비활성화 (관리자)",
        "tags": [
          "users"
        ]
      }
    },
    "/api/users/{id}/reactivate": {
      "post": {
        "operationId": "UserController_reactivate",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "사용자 ID",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "사용자 재활성화 성공"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "사용자 재활성화 (관리자)",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "SSOO Common API",
    "description": "Common module API reference",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
      "ApiSuccess": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              true
            ],
            "example": true
          },
          "data": {
            "type": "object",
            "additionalProperties": true
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "data"
        ]
      },
...
      "ApiErrorDetail": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "example": "INTERNAL_ERROR"
          },
          "message": {
            "type": "string",
            "example": "Internal server error"
          },
          "path": {
            "type": "string",
            "example": "/api/example"
          }
        },
        "required": [
          "code",
          "message"
        ]
      },
      "ApiError": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "enum": [
              false
            ],
            "example": false
          },
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "success",
          "error"
        ]
      },
...
      "UserProfileDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "loginId": {
            "type": "string"
          },
          "userName": {
            "type": "object",
            "nullable": true
          },
          "displayName": {
            "type": "object",
            "nullable": true
          },
          "email": {
            "type": "object",
            "nullable": true,
            "format": "email"
          },
          "phone": {
            "type": "object",
            "nullable": true
          },
          "avatarUrl": {
            "type": "object",
            "nullable": true
          },
          "departmentCode": {
            "type": "object",
            "nullable": true
          },
          "positionCode": {
            "type": "object",
            "nullable": true
          },
          "roleCode": {
            "type": "string"
          },
          "lastLoginAt": {
            "type": "string",
            "nullable": true,
            "format": "date-time"
          }
        },
        "required": [
          "id",
          "loginId",
          "roleCode"
        ]
      },
      "UpdateOwnProfileDto": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "description": "이름"
          },
          "displayName": {
            "type": "string",
            "description": "표시명"
          },
          "email": {
            "type": "string",
            "description": "이메일"
          },
          "phone": {
            "type": "string",
            "description": "전화번호"
          },
          "departmentCode": {
            "type": "string",
            "description": "부서"
          },
          "positionCode": {
            "type": "string",
            "description": "직책"
          }
        }
      },
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "loginId": {
            "type": "string",
            "description": "로그인 ID",
            "example": "john.doe"
          },
          "password": {
            "type": "string",
            "description": "비밀번호",
            "example": "Password1!"
          },
          "userName": {
            "type": "string",
            "description": "사용자명",
            "example": "홍길동"
          },
          "displayName": {
            "type": "string",
            "description": "표시명",
            "example": "길동"
          },
          "email": {
            "type": "string",
            "description": "이메일",
            "example": "john@example.com"
          },
          "phone": {
            "type": "string",
            "description": "전화번호",
            "example": "010-1234-5678"
          },
          "roleCode": {
            "type": "string",
            "description": "역할 코드",
            "example": "user",
            "default": "user"
          },
          "departmentCode": {
            "type": "string",
            "description": "부서 코드",
            "example": "DEV"
          },
          "positionCode": {
            "type": "string",
            "description": "직급 코드",
            "example": "senior"
          },
          "employeeNumber": {
            "type": "string",
            "description": "사번",
            "example": "E240015"
          },
          "companyName": {
            "type": "string",
            "description": "외부 소속 회사명",
            "example": "Acme Corp"
          },
          "customerId": {
            "type": "string",
            "description": "고객사 ID",
            "example": "1001"
          },
          "primaryAffiliationType": {
            "type": "string",
            "description": "Primary affiliation type",
            "enum": [
              "internal",
              "external"
            ],
            "example": "internal"
          }
        },
        "required": [
          "loginId",
          "password",
          "userName",
          "email"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "userName": {
            "type": "string",
            "description": "사용자명",
            "example": "홍길동"
          },
          "displayName": {
            "type": "string",
            "description": "표시명",
            "example": "길동"
          },
          "email": {
            "type": "string",
            "description": "이메일",
            "example": "john@example.com"
          },
          "phone": {
            "type": "string",
            "description": "전화번호",
            "example": "010-1234-5678"
          },
          "roleCode": {
            "type": "string",
            "description": "역할 코드",
            "example": "user"
          },
          "departmentCode": {
            "type": "string",
            "description": "부서 코드",
            "example": "DEV"
          },
          "positionCode": {
            "type": "string",
            "description": "직급 코드",
            "example": "senior"
          },
          "employeeNumber": {
            "type": "string",
            "description": "사번",
            "example": "E240015"
          },
          "companyName": {
            "type": "string",
            "description": "외부 소속 회사명",
            "example": "Acme Corp"
          },
          "customerId": {
            "type": "string",
            "description": "고객사 ID",
            "example": "1001"
          },
          "primaryAffiliationType": {
            "type": "string",
            "description": "Primary affiliation type",
            "enum": [
              "internal",
              "external"
            ],
            "example": "internal"
          },
          "password": {
            "type": "string",
            "description": "비밀번호 (변경 시에만)",
            "example": "NewPass1!"
          },
          "isActive": {
            "type": "boolean",
            "description": "활성 상태",
            "example": true
          }
        }
      },
...
    }
  }
}